Class "de"

An S4 class for differential evolution algorithm

Objects from the Class

Objects can be created by calls to the de function.

Slots

call

an object of class "call" representing the matched call;

type

a character string specifying the type of genetic algorithm used;

lower

a vector providing for each decision variable the lower bounds of the search space in case of real-valued or permutation encoded optimisations. Formerly this slot was named min;

upper

a vector providing for each decision variable the upper bounds of the search space in case of real-valued or permutation encoded optimizations. Formerly this slot was named max;

names

a vector of character strings providing the names of decision variables (optional);

popSize

the population size;

iter

the actual (or final) iteration of DE search;

run

the number of consecutive generations without any improvement in the best fitness value before the DE is stopped;

maxiter

the maximum number of iterations to run before the DE search is halted;

suggestions

a matrix of user provided solutions and included in the initial population;

population

the current (or final) population;

elitism

the number of best fitness individuals to survive at each generation;

stepsize

the stepsize or weighting factor;

pcrossover

the crossover probability;

pmutation

the mutation probability;

optim

a logical specifying whether or not a local search using general-purpose optimisation algorithms should be used;

fitness

the values of fitness function for the current (or final) population;

summary

a matrix of summary statistics for fitness values at each iteration (along the rows);

bestSol

if keepBest = TRUE, the best solutions at each iteration;

fitnessValue

the best fitness value at the final iteration;

solution

the value(s) of the decision variables giving the best fitness at the final iteration.

Author

Luca Scrucca

See also