Crossover operators in genetic algorithms

Functions implementing crossover genetic operator.

ga_spCrossover(object, parents, …)

gabin_spCrossover(object, parents, …)
gabin_uCrossover(object, parents, …)

gareal_spCrossover(object, parents, …)
gareal_waCrossover(object, parents, …)
gareal_laCrossover(object, parents, …)
gareal_blxCrossover(object, parents, …)
gareal_laplaceCrossover(object, parents, a = 0, b = 0.15, …)

gaperm_cxCrossover(object, parents, …)
gaperm_pmxCrossover(object, parents, …)
gaperm_oxCrossover(object, parents, …)
gaperm_pbxCrossover(object, parents, …)

Arguments

object
An object of class "ga", usually resulting from a call to function ga.

parents
A two-rows matrix of values indexing the parents from the current population.

…
Further arguments passed to or from other methods.

a, b
Parameters of genetic operators.

Value

Return a list with two elements:

children

a matrix of dimension 2 times the number of decision variables containing the generated offsprings;

fitness

a vector of length 2 containing the fitness values for the offsprings. A value NA is returned if an offspring is different (which is usually the case) from the two parents.

See also

ga