Selection operators in genetic algorithms

Functions implementing selection genetic operator.

ga_lrSelection(object, r = 2/(object@popSize * (object@popSize - 1)),
                       q = 2/object@popSize, ...)
ga_nlrSelection(object, q = 0.25, ...)
ga_rwSelection(object, ...)
ga_tourSelection(object, k = 3, ...)

gabin_lrSelection(object, r = 2/(object@popSize * (object@popSize - 1)), 
                          q = 2/object@popSize, ...)
gabin_nlrSelection(object, q = 0.25, ...)
gabin_rwSelection(object, ...)
gabin_tourSelection(object, k = 3, ...)

gareal_lrSelection(object, r = 2/(object@popSize * (object@popSize - 1)), 
                           q = 2/object@popSize, ...)
gareal_nlrSelection(object, q = 0.25, ...)
gareal_rwSelection(object, ...)
gareal_tourSelection(object, k = 3, ...)
gareal_lsSelection(object, ...)
gareal_sigmaSelection(object, ...)

gaperm_lrSelection(object, r = 2/(object@popSize * (object@popSize - 1)), 
                           q = 2/object@popSize, ...)
gaperm_nlrSelection(object, q = 0.25, ...)
gaperm_rwSelection(object, ...)
gaperm_tourSelection(object, k = 3, ...)

gareal_de(object, F = 0.8, p = 0.5, ...)

Arguments

object

An object of class "ga", usually resulting from a call to function ga.

r

A tuning parameter for the GA selection operator.

q

A tuning parameter for the GA selection operator.

k

A tuning parameter for the GA selection operator.

F, p

Tuning parameters for the DE selection operator.

...

Further arguments passed to or from other methods.

Value

Return a list with two elements:

population

a matrix of dimension object@popSize times the number of decision variables containing the selected individuals or strings;

fitness

a vector of length object@popSize containing the fitness values for the selected individuals.

Author

Luca Scrucca

See also