Plot of Differential Evolution search path

The plot method for de-class objects gives a plot of best and average fitness values found during the iterations of the DE search.

# S4 method for de
plot(x, y, ylim, cex.points = 0.7, 
     col = c("green3", "dodgerblue3",  adjustcolor("green3", alpha.f = 0.1)),
     pch = c(16, 1), lty = c(1,2), legend = TRUE, grid = graphics:::grid, ...)

Arguments

x
An object of class "ga".

y
Not used.

ylim
A vector of two values specifying the limits on the y-axis.

cex.points
The magnification to be used for points.

col
The colours to be used for best and average fitness values.

pch
The type of points to be used for best and average fitness values.

lty
The type of lines to be used for best and average fitness values.

legend
A logical specifying if a legend should be included.

grid
A function for grid drawing of NULL to avoid drawing one.

...
Further arguments, currently not used.

Details

Plot best and average fitness values at each iteration of DE search.

Value

The method invisibly return a data.frame with the iterations and summary statistics for the fitness function evaluated at each iteration.

See also

Author

Luca Scrucca

Examples

# See examples in help(de)