Plot of Islands Genetic Algorithm search path

The plot method for gaisl-class objects gives a plot of best fitness values found in each island during the GA iterations.

# S4 method for gaisl
plot(x, y, ...)

Arguments

x
An object of class "gaisl".

y
Not used.

...
Further arguments passed to plot.default, such as ylim, ylab, etc., or to matplot, such as col, lty, and lwd.

Details

Plot best fitness values found in each island during the GA iterations.

Value

The method invisibly return a list with the following components:

iter

a vector of values specifying the iteration.

summary

a matrix of best fitness values for each island along the columns.

See also

Author

Luca Scrucca

Examples

# See examples in help(gaisl)