Summary for Genetic Algorithms

Summary method for class ga-class.

# S4 method for ga
summary(object, …)

# S3 method for summary.ga
print(x, digits = getOption("digits"), …)

Arguments

object
an object of class ga-class.

x
an object of class summary.ga.

digits
number of significant digits.

…
further arguments passed to or from other methods.

Value

The summary function returns an object of class summary.ga which can be printed by the corresponding print method. The function also returns invisibly a list with the information from the genetic algorithm search.

See also

ga

Examples

f <- function(x)  abs(x)+cos(x)
GA <- ga(type = "real-valued",
         fitness = function(x) -f(x),
         lower = -20, upper = 20, run = 50)

#> GA | iter = 1 | Mean = -10.223093 | Best =  -1.023586
#> GA | iter = 2 | Mean = -6.929785 | Best = -1.023586
#> GA | iter = 3 | Mean = -4.140516 | Best = -1.000488
#> GA | iter = 4 | Mean = -2.354117 | Best = -1.000488
#> GA | iter = 5 | Mean = -2.688997 | Best = -1.000488
#> GA | iter = 6 | Mean = -2.495615 | Best = -1.000488
#> GA | iter = 7 | Mean = -2.473460 | Best = -1.000488
#> GA | iter = 8 | Mean = -2.669237 | Best = -1.000488
#> GA | iter = 9 | Mean = -2.500355 | Best = -1.000488
#> GA | iter = 10 | Mean = -2.213886 | Best = -1.000488
#> GA | iter = 11 | Mean = -1.791104 | Best = -1.000044
#> GA | iter = 12 | Mean = -2.585730 | Best = -1.000044
#> GA | iter = 13 | Mean = -1.540502 | Best = -1.000044
#> GA | iter = 14 | Mean = -2.471441 | Best = -1.000044
#> GA | iter = 15 | Mean = -1.437884 | Best = -1.000044
#> GA | iter = 16 | Mean = -2.657705 | Best = -1.000044
#> GA | iter = 17 | Mean = -1.569611 | Best = -1.000044
#> GA | iter = 18 | Mean = -2.153061 | Best = -1.000044
#> GA | iter = 19 | Mean = -2.199311 | Best = -1.000044
#> GA | iter = 20 | Mean = -3.245032 | Best = -1.000044
#> GA | iter = 21 | Mean = -3.673713 | Best = -1.000044
#> GA | iter = 22 | Mean = -2.571690 | Best = -1.000044
#> GA | iter = 23 | Mean = -2.904425 | Best = -1.000044
#> GA | iter = 24 | Mean = -1.778804 | Best = -1.000044
#> GA | iter = 25 | Mean = -2.614047 | Best = -1.000044
#> GA | iter = 26 | Mean = -2.108911 | Best = -1.000044
#> GA | iter = 27 | Mean = -1.696283 | Best = -1.000044
#> GA | iter = 28 | Mean = -1.470633 | Best = -1.000044
#> GA | iter = 29 | Mean = -2.664023 | Best = -1.000044
#> GA | iter = 30 | Mean = -1.889151 | Best = -1.000003
#> GA | iter = 31 | Mean = -2.287573 | Best = -1.000003
#> GA | iter = 32 | Mean = -2.171350 | Best = -1.000003
#> GA | iter = 33 | Mean = -2.440384 | Best = -1.000003
#> GA | iter = 34 | Mean = -2.447518 | Best = -1.000003
#> GA | iter = 35 | Mean = -2.132850 | Best = -1.000003
#> GA | iter = 36 | Mean = -1.766190 | Best = -1.000003
#> GA | iter = 37 | Mean = -1.400201 | Best = -1.000003
#> GA | iter = 38 | Mean = -2.424358 | Best = -1.000003
#> GA | iter = 39 | Mean = -2.331598 | Best = -1.000003
#> GA | iter = 40 | Mean = -1.670601 | Best = -1.000003
#> GA | iter = 41 | Mean = -2.827666 | Best = -1.000003
#> GA | iter = 42 | Mean = -2.373023 | Best = -1.000003
#> GA | iter = 43 | Mean = -2.101009 | Best = -1.000003
#> GA | iter = 44 | Mean = -1.634584 | Best = -1.000003
#> GA | iter = 45 | Mean = -2.560502 | Best = -1.000003
#> GA | iter = 46 | Mean = -2.041194 | Best = -1.000003
#> GA | iter = 47 | Mean = -2.096030 | Best = -1.000003
#> GA | iter = 48 | Mean = -2.853282 | Best = -1.000003
#> GA | iter = 49 | Mean = -2.067944 | Best = -1.000003
#> GA | iter = 50 | Mean = -2.076934 | Best = -1.000003
#> GA | iter = 51 | Mean = -3.221209 | Best = -1.000003
#> GA | iter = 52 | Mean = -3.174651 | Best = -1.000003
#> GA | iter = 53 | Mean = -2.113756 | Best = -1.000003
#> GA | iter = 54 | Mean = -2.517285 | Best = -1.000003
#> GA | iter = 55 | Mean = -1.894761 | Best = -1.000003
#> GA | iter = 56 | Mean = -2.341173 | Best = -1.000003
#> GA | iter = 57 | Mean = -2.198096 | Best = -1.000003
#> GA | iter = 58 | Mean = -1.877188 | Best = -1.000003
#> GA | iter = 59 | Mean = -1.814751 | Best = -1.000000
#> GA | iter = 60 | Mean = -1.358631 | Best = -1.000000
#> GA | iter = 61 | Mean = -1.966156 | Best = -1.000000
#> GA | iter = 62 | Mean = -2.147889 | Best = -1.000000
#> GA | iter = 63 | Mean = -1.860529 | Best = -1.000000
#> GA | iter = 64 | Mean = -1.977459 | Best = -1.000000
#> GA | iter = 65 | Mean = -2.136353 | Best = -1.000000
#> GA | iter = 66 | Mean = -2.475267 | Best = -1.000000
#> GA | iter = 67 | Mean = -1.525329 | Best = -1.000000
#> GA | iter = 68 | Mean = -3.018618 | Best = -1.000000
#> GA | iter = 69 | Mean = -1.748411 | Best = -1.000000
#> GA | iter = 70 | Mean = -2.658537 | Best = -1.000000
#> GA | iter = 71 | Mean = -3.146135 | Best = -1.000000
#> GA | iter = 72 | Mean = -1.773501 | Best = -1.000000
#> GA | iter = 73 | Mean = -2.287908 | Best = -1.000000
#> GA | iter = 74 | Mean = -2.21711 | Best = -1.00000
#> GA | iter = 75 | Mean = -2.387238 | Best = -1.000000
#> GA | iter = 76 | Mean = -2.519909 | Best = -1.000000
#> GA | iter = 77 | Mean = -2.448445 | Best = -1.000000
#> GA | iter = 78 | Mean = -2.018379 | Best = -1.000000
#> GA | iter = 79 | Mean = -2.208071 | Best = -1.000000
#> GA | iter = 80 | Mean = -2.241581 | Best = -1.000000
#> GA | iter = 81 | Mean = -2.271686 | Best = -1.000000
#> GA | iter = 82 | Mean = -1.94703 | Best = -1.00000
#> GA | iter = 83 | Mean = -1.980435 | Best = -1.000000
#> GA | iter = 84 | Mean = -2.459314 | Best = -1.000000
#> GA | iter = 85 | Mean = -2.596734 | Best = -1.000000
#> GA | iter = 86 | Mean = -1.883449 | Best = -1.000000
#> GA | iter = 87 | Mean = -1.615717 | Best = -1.000000
#> GA | iter = 88 | Mean = -1.847583 | Best = -1.000000
#> GA | iter = 89 | Mean = -1.891623 | Best = -1.000000
#> GA | iter = 90 | Mean = -2.622436 | Best = -1.000000
#> GA | iter = 91 | Mean = -2.260501 | Best = -1.000000
#> GA | iter = 92 | Mean = -2.212829 | Best = -1.000000
#> GA | iter = 93 | Mean = -2.233524 | Best = -1.000000
#> GA | iter = 94 | Mean = -1.995075 | Best = -1.000000
#> GA | iter = 95 | Mean = -2.771056 | Best = -1.000000
#> GA | iter = 96 | Mean = -2.181416 | Best = -1.000000
#> GA | iter = 97 | Mean = -3.342357 | Best = -1.000000
#> GA | iter = 98 | Mean = -1.984024 | Best = -1.000000
#> GA | iter = 99 | Mean = -2.388593 | Best = -1.000000
#> GA | iter = 100 | Mean = -1.693897 | Best = -1.000000

out <- summary(GA)
print(out)

#> ── Genetic Algorithm ─────────────────────────────────────────────────────────────── 
#> 
#> GA settings: 
#> Type                  =  real-valued 
#> Population size       =  50 
#> Number of generations =  100 
#> Elitism               =  2 
#> Crossover probability =  0.8 
#> Mutation probability  =  0.1 
#> Search domain = 
#>        x1
#> lower -20
#> upper  20
#> 
#> GA results: 
#> Iterations             = 100 
#> Fitness function value = -1 
#> Solution = 
#>                    x1
#> [1,] -0.0000002104444

str(out)

#> List of 11
#>  $ type       : chr "real-valued"
#>  $ popSize    : num 50
#>  $ maxiter    : num 100
#>  $ elitism    : num 2
#>  $ pcrossover : num 0.8
#>  $ pmutation  : num 0.1
#>  $ domain     : num [1:2, 1] -20 20
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "lower" "upper"
#>   .. ..$ : chr "x1"
#>  $ suggestions: NULL
#>  $ iter       : int 100
#>  $ fitness    : num -1
#>  $ solution   : num [1, 1] -0.00000021
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr "x1"
#>  - attr(*, "class")= chr "summary.ga"