Summary for Islands Genetic Algorithms

Summary method for class gaisl-class.

# S4 method for gaisl
summary(object, …)

# S3 method for summary.gaisl
print(x, digits = getOption("digits"), …)

Arguments

object
an object of class gaisl-class.

x
an object of class summary.gaisl.

digits
number of significant digits.

…
further arguments passed to or from other methods.

Value

The summary function returns an object of class summary.gaisl which can be printed by the corresponding print method. The function also returns invisibly a list with the information from the islands genetic algorithm search.

See also

gaisl

Examples

f <- function(x)  abs(x)+cos(x)
GA <- gaisl(type = "real-valued",
            fitness = function(x) -f(x),
            lower = -20, upper = 20, run = 10,
            numIslands = 4, parallel = 2) # CRAN check limits ncores = 2

#> Islands GA | epoch = 1 
#> Mean = -2.533351 | Best = -1.044710 
#> Mean = -3.240238 | Best = -1.011332 
#> Mean = -2.129411 | Best = -1.000026 
#> Mean = -2.103206 | Best = -1.011947 
#> Islands GA | epoch = 2 
#> Mean = -3.002379 | Best = -1.000922 
#> Mean = -1.836892 | Best = -1.002404 
#> Mean = -3.201009 | Best = -1.000026 
#> Mean = -2.096590 | Best = -1.000026 
#> Islands GA | epoch = 3 
#> Mean = -2.197200 | Best = -1.000026 
#> Mean = -1.155922 | Best = -1.000922 
#> Mean = -2.251763 | Best = -1.000026 
#> Mean = -1.799951 | Best = -1.000026 
#> Islands GA | epoch = 4 
#> Mean = -1.674199 | Best = -1.000026 
#> Mean = -1.963850 | Best = -1.000026 
#> Mean = -2.498293 | Best = -1.000026 
#> Mean = -1.532814 | Best = -1.000003 
#> Islands GA | epoch = 5 
#> Mean = -1.806703 | Best = -1.000003 
#> Mean = -2.745763 | Best = -1.000026 
#> Mean = -2.455349 | Best = -1.000026 
#> Mean = -2.189915 | Best = -1.000003 
#> Islands GA | epoch = 6 
#> Mean = -2.188936 | Best = -1.000003 
#> Mean = -1.311417 | Best = -1.000003 
#> Mean = -1.615910 | Best = -1.000026 
#> Mean = -2.156924 | Best = -1.000003 
#> Islands GA | epoch = 7 
#> Mean = -1.391761 | Best = -1.000003 
#> Mean = -1.751078 | Best = -1.000003 
#> Mean = -1.218731 | Best = -1.000003 
#> Mean = -1.437053 | Best = -1.000003 
#> Islands GA | epoch = 8 
#> Mean = -1.117236 | Best = -1.000003 
#> Mean = -1.855110 | Best = -1.000003 
#> Mean = -1.251064 | Best = -1.000003 
#> Mean = -2.228851 | Best = -1.000003 

out <- summary(GA)
print(out)

#> ── Islands Genetic Algorithm ─────────── 
#> 
#> GA settings: 
#> Type                  =  real-valued 
#> Number of islands     =  4 
#> Islands pop. size     =  25 
#> Migration rate        =  0.1 
#> Migration interval    =  10 
#> Elitism               =  1 
#> Crossover probability =  0.8 
#> Mutation probability  =  0.1 
#> Search domain = 
#>        x1
#> lower -20
#> upper  20
#> 
#> GA results: 
#> Iterations              = 80 
#> Epochs                  = 8 
#> Fitness function values = -1.000003 -1.000003 -1.000003 -1.000003 
#> Solutions = 
#>                  x1
#> [1,] 0.000002570573
#> [2,] 0.000002570573
#> [3,] 0.000002570573
#> [4,] 0.000002570573

str(out)

#> List of 13
#>  $ type             : chr "real-valued"
#>  $ numIslands       : num 4
#>  $ popSizeIslands   : num 25
#>  $ migrationRate    : num 0.1
#>  $ migrationInterval: num 10
#>  $ elitism          : num 1
#>  $ pcrossover       : num 0.8
#>  $ pmutation        : num 0.1
#>  $ domain           : num [1:2, 1] -20 20
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:2] "lower" "upper"
#>   .. ..$ : chr "x1"
#>  $ iter             : int 80
#>  $ epoch            : num 8
#>  $ fitnessValues    : num [1:4] -1 -1 -1 -1
#>  $ solutions        : num [1:4, 1] 0.00000257 0.00000257 0.00000257 0.00000257
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr "x1"
#>  - attr(*, "class")= chr "summary.gaisl"