Class "gaisl"

An S4 class for islands genetic algorithms (ISLGAs)

Objects from the Class

Objects can be created by calls to the gaisl function.

Slots

call

an object of class "call" representing the matched call;

type

a character string specifying the type of genetic algorithm used;

lower

a vector providing for each decision variable the lower bounds of the search space in case of real-valued or permutation encoded optimisations. Formerly this slot was named min;

upper

a vector providing for each decision variable the upper bounds of the search space in case of real-valued or permutation encoded optimizations. Formerly this slot was named max;

nBits

a value specifying the number of bits to be used in binary encoded optimizations;

names

a vector of character strings providing the names of decision variables (optional);

popSize

the population size;

numIslands

the number of islands;

migrationRate

the migration rate;

migrationInterval

the migration interval;

maxiter

the maximum number of ISLGA iterations before the search is halted;

run

the number of consecutive generations without any improvement in the best fitness value before the ISLGA is stopped;

maxiter

the maximum number of iterations to run before the GA search is halted;

suggestions

a matrix of user provided solutions and included in the initial population;

elitism

the number of best fitness individuals to survive at each generation;

pcrossover

the crossover probability;

pmutation

the mutation probability;

islands

a list containing the "ga" objects corresponding to each island GA evolution;

summary

a list of matrices of summary statistics for fitness values at each iteration (along the rows). Each element of the list corresponds to the evolution of an island;

fitnessValues

a list of best fitness values found in each island at the final iteration;

solutions

a list of matrices, one for each island, containing the values of the decision variables giving the best fitness at the final iteration;

fitnessValue

the best fitness value at the final iteration;

solution

a matrix containing the values of the decision variables giving the best fitness at the final iteration.

See also

For examples of usage see gaisl.