Summary for Islands Genetic Algorithms

Summary method for class gaisl-class.

# S4 method for gaisl
summary(object, …)

# S3 method for summary.gaisl
print(x, digits = getOption("digits"), …)

Arguments

object
an object of class gaisl-class.

x
an object of class summary.gaisl.

digits
number of significant digits.

…
further arguments passed to or from other methods.

Value

The summary function returns an object of class summary.gaisl which can be printed by the corresponding print method. The function also returns invisibly a list with the information from the islands genetic algorithm search.

See also

Examples

# NOT RUN {
f <- function(x)  abs(x)+cos(x)
GA <- gaisl(type = "real-valued",
            fitness = function(x) -f(x),
            lower = -20, upper = 20, run = 10,
            numIslands = 4)
out <- summary(GA)
print(out)
str(out)
# }